+++ src/lcd/jsupm_i2clcd.i
@@ -9,9 +9,9 @@
 }
 
 %include "ssd.h"
-%include "i2clcd.h"
+%include "lcd.h"
 %{
-    #include "i2clcd.h"
+    #include "lcd.h"
 %}
 
 %include "jhd1313m1.h"
