+++ src/lcd/lcm1602.h
@@ -28,7 +28,7 @@
 #pragma once
 
 #include <string>
-#include "i2clcd.h"
+#include "lcd.h"
 
 namespace upm
 {
