+++ src/lcd/ssd1308.h
@@ -25,7 +25,7 @@
 #pragma once
 
 #include <string>
-#include "i2clcd.h"
+#include "lcd.h"
 #include "ssd.h"
 
 namespace upm
